refactor: rearrange file
