--- packages/mirror-tv-next/app/(amp-layout)/story/amp/[slug]/page.tsx
@@ -1,6 +1,0 @@
-export const runtime = 'edge'
-export const dynamic = 'force-static'
-
-export default function AmpPage() {
-  return <div>amp page</div>
-}
