+++ packages/mirror-tv-next/app/anchorperson/[slug]/page.tsx
