+++ packages/mirror-tv-next/pages/layout.tsx
@@ -3,8 +3,7 @@ import styles from '~/components/story/amp/_styles/layout.module.scss'
 import { Noto_Sans } from 'next/font/google'
 import Head from 'next/head'
 
-export const runtime = 'edge'
-export const dynamic = 'force-static'
+export const config = { amp: true }
 
 const noto_sans = Noto_Sans({
   weight: ['400', '500', '600'],
